Join SimulationContext and BatchContext
