+++ src/ert/simulator/__init__.py
@@ -1,9 +1,7 @@
 from .batch_simulator import BatchSimulator
 from .batch_simulator_context import BatchContext
-from .simulation_context import SimulationContext
 
 __all__ = [
     "BatchSimulator",
     "BatchContext",
-    "SimulationContext",
 ]
